refactor: renamed Summary -> ReviewSwap
